requirements.txt
django
django-ninja
django-ninja-simple-jwt
pillow
mysqlclient # perlu install mysql-client lewat brew
